refactored functions so it resembles that of a traditional express app
